Division and remainder implementation

+++ CG.hpp
@@ -1,8 +1,8 @@
-int dataOffset=0;
-int dataLocation(int type){
-	int temp=dataOffset;
+long long dataOffset=0;
+long long dataLocation(int type){
+	long long temp=dataOffset;
 	dataOffset+=type;
 	return temp;
 }
 
-int codeOffset=0;
+long long codeOffset=0;
