+++ ST.hpp
@@ -18,7 +18,7 @@ bool addVar(std::string name){
 	stable[name]=a;
 	return true;
 }
-int getVar(std::string name){
+long long getVar(std::string name){
 	it=stable.find(name);
 	if(it!=stable.end()){
 		return it->second.offset;
